+++ src/context/datasetContext/index.tsx
@@ -27,9 +27,10 @@ import {
 } from "@inrupt/solid-client";
 
 import useDataset from "../../hooks/useDataset";
+import { WithAcp } from "@inrupt/solid-client/acp/acp";
 
 export interface IDatasetContext {
-  solidDataset: SolidDataset | (SolidDataset & WithResourceInfo) | undefined;
+  solidDataset: SolidDataset | (SolidDataset & WithResourceInfo) | (SolidDataset & WithAcp) | undefined;
   setDataset: (solidDataset: SolidDataset) => void;
 }
 
@@ -48,7 +49,7 @@ export interface IDatasetProvider {
   loading?: React.ReactNode;
   loadingComponent?: React.ComponentType | null;
   onError?(error: Error): void | null;
-  solidDataset?: SolidDataset | (SolidDataset & WithResourceInfo);
+  solidDataset?: SolidDataset | (SolidDataset & WithResourceInfo) | (SolidDataset & WithAcp);
   datasetUrl?: UrlString | string;
 }
 
@@ -81,7 +82,7 @@ export const DatasetProvider = ({
 
   // Provide a setDataset function so that child components can update.
   const [stateDataset, setDataset] = useState<
-    SolidDataset | (SolidDataset & WithResourceInfo) | undefined
+    SolidDataset | (SolidDataset & WithResourceInfo) | (SolidDataset & WithAcp)| undefined
   >(datasetToUse);
 
   // If the dataset is asynchronously loaded, make sure to set the new state value.